# swagger-docs.json
{"openapi":"3.0.0","info":{"title":"DevApi Api","description":"DevApi Api","version":"0.0.1","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"RegisterDTO":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"UserInfo":{"type":"object","properties":{"_id":{"type":"string"},"email":{"type":"string"},"roles":{"type":"array","items":{"type":"string"}}},"required":["_id","email","roles"]},"AuthenticatedUserDTO":{"type":"object","properties":{"user":{"$ref":"#/components/schemas/UserInfo"},"token":{"type":"string"}},"required":["user","token"]},"LoginDTO":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"ObjectID":{"type":"object","properties":{}},"User":{"type":"object","properties":{"_id":{"$ref":"#/components/schemas/ObjectID"},"email":{"type":"string"},"password":{"type":"string","nullable":true},"roles":{"type":"array","items":{"type":"string"}},"createdAt":{"format":"date-time","type":"string"}},"required":["_id","email","password","roles","createdAt"]},"PaginationInfos":{"type":"object","properties":{"total":{"type":"number"},"skip":{"type":"number"},"limit":{"type":"number"}},"required":["total","skip","limit"]},"Connector":{"type":"object","properties":{"_id":{"type":"string"},"name":{"type":"string"},"type":{"type":"string","enum":["REST","BD","SOAP"]},"privacy":{"type":"string","enum":["PUBLIC","PRIVATE"]},"baseUrl":{"type":"string"},"logoUrl":{"type":"string"},"category":{"type":"string"},"description":{"type":"string"},"status":{"type":"string"},"createdAt":{"format":"date-time","type":"string"}},"required":["_id","name","type","privacy","baseUrl","logoUrl","category","description","status","createdAt"]},"ListConnectors":{"type":"object","properties":{"pagination":{"$ref":"#/components/schemas/PaginationInfos"},"results":{"type":"array","items":{"$ref":"#/components/schemas/Connector"}}},"required":["pagination","results"]},"CreateConnectorDTO":{"type":"object","properties":{"name":{"type":"string"},"type":{"type":"string","enum":["REST","BD","SOAP"]},"privacy":{"type":"string","enum":["PUBLIC","PRIVATE"]},"baseUrl":{"type":"string"},"logoUrl":{"type":"string"},"category":{"type":"string"},"description":{"type":"string"},"status":{"type":"string"}},"required":["name","type","privacy","baseUrl","logoUrl","category","description","status"]},"UpdateConnectorDTO":{"type":"object","properties":{"name":{"type":"string"},"type":{"type":"string","enum":["REST","BD","SOAP"]},"privacy":{"type":"string","enum":["PUBLIC","PRIVATE"]},"baseUrl":{"type":"string"},"logoUrl":{"type":"string"},"category":{"type":"string"},"description":{"type":"string"},"status":{"type":"string"}},"required":["name","type","privacy","baseUrl","logoUrl","category","description","status"]},"PatchConnectorDTO":{"type":"object","properties":{"name":{"type":"string"},"type":{"type":"string","enum":["REST","BD","SOAP"]},"privacy":{"type":"string","enum":["PUBLIC","PRIVATE"]},"baseUrl":{"type":"string"},"logoUrl":{"type":"string"},"category":{"type":"string"},"description":{"type":"string"},"status":{"type":"string"}}}}},"paths":{"/authentications/users":{"post":{"operationId":"AuthenticationsController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDTO"}}}},"responses":{"201":{"description":"Usuário cadastrado com sucesso.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthenticatedUserDTO"}}}}},"tags":["authentications"]}},"/authentications/admins":{"post":{"operationId":"AuthenticationsController_registerAdmin","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDTO"}}}},"responses":{"201":{"description":"Administrador cadastrado com sucesso.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthenticatedUserDTO"}}}}},"tags":["authentications"]}},"/authentications/login":{"post":{"operationId":"AuthenticationsController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDTO"}}}},"responses":{"200":{"description":"Login realizado com sucesso.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthenticatedUserDTO"}}}}},"tags":["authentications"]}},"/authentications/me":{"get":{"operationId":"AuthenticationsController_me","parameters":[],"responses":{"201":{"description":"Usuário obtido com sucesso.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"tags":["authentications"],"security":[{"bearer":[]}]}},"/connectors":{"get":{"operationId":"ConnectorsController_getMany","parameters":[{"name":"name","required":false,"in":"query","schema":{"type":"string"}},{"name":"type","required":false,"in":"query","schema":{"enum":["REST","BD","SOAP"],"type":"string"}},{"name":"privacy","required":false,"in":"query","schema":{"enum":["PUBLIC","PRIVATE"],"type":"string"}},{"name":"baseUrl","required":false,"in":"query","schema":{"type":"string"}},{"name":"logoUrl","required":false,"in":"query","schema":{"type":"string"}},{"name":"category","required":false,"in":"query","schema":{"type":"string"}},{"name":"description","required":false,"in":"query","schema":{"type":"string"}},{"name":"status","required":false,"in":"query","schema":{"type":"string"}},{"name":"skip","required":false,"in":"query","schema":{"default":0,"type":"number"}},{"name":"limit","required":false,"in":"query","schema":{"default":10,"type":"number"}}],"responses":{"200":{"description":"Conectores obtidos com sucesso","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ListConnectors"}}}}},"tags":["connectors"]},"post":{"operationId":"ConnectorsController_store","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateConnectorDTO"}}}},"responses":{"201":{"description":"Conector cadastrada com sucesso.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Connector"}}}}},"tags":["connectors"],"security":[{"bearer":[]}]}},"/connectors/{id}":{"get":{"operationId":"ConnectorsController_getOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Conector obtido com sucesso.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Connector"}}}}},"tags":["connectors"]},"put":{"operationId":"ConnectorsController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateConnectorDTO"}}}},"responses":{"200":{"description":"Conector atualizado com sucesso."}},"tags":["connectors"],"security":[{"bearer":[]}]},"patch":{"operationId":"ConnectorsController_partialUpdate","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PatchConnectorDTO"}}}},"responses":{"200":{"description":"Conector atualizado com sucesso."}},"tags":["connectors"],"security":[{"bearer":[]}]}}}}